+++ src/components/SFormField/SFormField.vue
@@ -1,0 +1,11 @@
+<template>
+  <div class="form-field">
+    <slot />
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'SFormField'
+}
+</script>