+++ src/components/SFormField/index.js
@@ -1,0 +1,7 @@
+import SFormField from './SFormField.vue'
+
+export { SFormField }
+
+export default {
+  install: Vue => Vue.component(SFormField.name, SFormField)
+}