+++ src/components/SFormValidation/index.js
@@ -1,0 +1,7 @@
+import SFormValidation from './SFormValidation.vue'
+
+export { SFormValidation }
+
+export default {
+  install: Vue => Vue.component(SFormValidation.name, SFormValidation)
+}